Update unsupported target error to reflect addition of ESP32 support (#22)
Support for the ESP32 family of microcontrollers was recently added to the extension.

When a user attempts a filesystem upload while a board of an unsupported architecture is selected in Arduino IDE, the
extension displays a helpful error message which includes a list of supported targets. This message was not updated at
the time the ESP32 support was added.