Fix octothorpe comments in ESP32 partitions
The code took each partitions.csv line and cut off everything after
a "#"....and then threq out the cut line and parsed the original.
This means it would parse commented lines as if they just had a
name of "#spiffs", for example.

Actually use the trimmed line in calculations.

Fixes #90